requirements.txt
python-dotenv
transformers[sentencepiece]
torch
sacremoses
torch-model-archiver
torchserve